feat(Lotto): Change package and add class validation
Added validation for distinct number among lotto numbers
